Moved Mesh tooling to its own file
