+++ test/runtests.jl
@@ -9,6 +9,10 @@ end
   include("DualSimplicialSets.jl")
 end
 
+@testset "Utility Functions" begin
+  include("MeshUtils.jl")
+end
+
 @testset "Visualization" begin
   include("Visualization.jl")
 end
